feat(worker,web): audit worker and rename OutputScreen to WorkspaceScreen
