+++ apps/server/src/socket.ts
@@ -9,10 +9,20 @@ import { env } from './env'
 const pubClient = new Redis(process.env.REDIS_URL || 'redis://localhost:6379', {
     maxRetriesPerRequest: null,
 })
-const subClient = pubClient.duplicate()
+pubClient.on('error', (err) => {
+    console.error('[Socket Redis PubClient Error]', err?.message || err)
+})
+
+const subClient = pubClient.duplicate({ enableReadyCheck: false })
+subClient.on('error', (err) => {
+    console.error('[Socket Redis SubClient Error]', err?.message || err)
+})
 
 // this one is specifically for subscribing to worker session events
-const redisSubClient = pubClient.duplicate()
+const redisSubClient = pubClient.duplicate({ enableReadyCheck: false })
+redisSubClient.on('error', (err) => {
+    console.error('[Socket Redis EventSubClient Error]', err?.message || err)
+})
 
 let io: Server
 
@@ -137,25 +147,32 @@ export function initSocket(httpServer: any) {
             'send_prompt',
             async (data: { sessionId: string; prompt: string; projectId: string }) => {
                 try {
+                    console.log(
+                        `[Core Socket] Received prompt for session ${data.sessionId} from user ${socket.data.userId || 'anonymous'}: "${data.prompt?.slice(0, 60)}..."`
+                    )
+
                     // fetch user secrets (phase 3.6 secrets management)
-                    // we'll mock decryption here since decryption logic belongs to secret service
-                    const secrets: any[] = [] // secret model doesn't exist yet, passing empty secrets
+                    const secrets: any[] = [] // secret model fallback
                     const decryptedSecrets = secrets.map((s: any) => ({
                         key: s.key,
-                        value: s.value, // assuming decrypted or decryption utility here
+                        value: s.value,
                     }))
 
                     // enqueue to worker
                     const agentJobsQueue = new Queue('agent_jobs', { connection: pubClient as any })
-                    await agentJobsQueue.add('run_agent', {
+                    const job = await agentJobsQueue.add('run_agent', {
                         sessionId: data.sessionId,
                         projectId: data.projectId,
                         userId: socket.data.userId,
                         prompt: data.prompt,
-                        secrets: decryptedSecrets, // injecting decrypted secrets into payload
+                        secrets: decryptedSecrets,
                     })
+
+                    console.log(
+                        `[Core Socket] Successfully enqueued job ${job.id} to BullMQ 'agent_jobs' for session ${data.sessionId}`
+                    )
                 } catch (err: any) {
-                    console.error('[Socket] Failed to enqueue agent job:', err)
+                    console.error('[Core Socket] Failed to enqueue agent job:', err)
                     socket.emit('error', { message: 'Failed to start agent: ' + err.message })
                 }
             }
