+++ apps/web/src/features/preview/components/OutputHeader.tsx
@@ -1,59 +1,4 @@
-import React from 'react'
+import { WorkspaceHeader } from './WorkspaceHeader'
 
-import { OutputHeaderActions } from './OutputHeaderActions'
-import { OutputHeaderCenterControls } from './OutputHeaderCenterControls'
-import { OutputHeaderViewTabs } from './OutputHeaderViewTabs'
-
-import type { OutputHeaderProps } from '@/features/preview/types'
-
-export const OutputHeader: React.FC<OutputHeaderProps> = ({
-    activeTab,
-    setActiveTab,
-    device,
-    setDevice,
-    isSidebarCollapsed,
-    onToggleSidebar,
-    onOpenNewTab,
-    onBack,
-    projectName,
-    projectId,
-    versions,
-    activeVersionId,
-    isVersionLoading,
-    onSelectVersion,
-    onDownload,
-    selectedModel,
-    setSelectedModel,
-    onRefresh,
-}) => {
-    return (
-        <header className="h-12 flex items-center justify-between px-3 bg-[#141414] backdrop-blur-sm shrink-0 z-[45] gap-3">
-            <OutputHeaderViewTabs
-                activeTab={activeTab}
-                setActiveTab={setActiveTab}
-                isSidebarCollapsed={isSidebarCollapsed}
-                onToggleSidebar={onToggleSidebar}
-                onBack={onBack}
-            />
-
-            <OutputHeaderCenterControls
-                device={device}
-                setDevice={setDevice}
-                isSidebarCollapsed={isSidebarCollapsed}
-                onToggleSidebar={onToggleSidebar}
-                onOpenNewTab={onOpenNewTab}
-                onRefresh={onRefresh}
-            />
-
-            <OutputHeaderActions
-                projectName={projectName}
-                projectId={projectId}
-                versions={versions}
-                activeVersionId={activeVersionId}
-                isVersionLoading={isVersionLoading}
-                onSelectVersion={onSelectVersion}
-                onDownload={onDownload}
-            />
-        </header>
-    )
-}
+export { WorkspaceHeader }
+export const OutputHeader = WorkspaceHeader
